requirements.txt
fastapi==0.111.0
python-telegram-bot==21.1.1
ujson==5.9.0